Update mcdrive
* refresh all dependencies
* switch Java version to 11
* remove old servlets and services
* change database credentials
* update database tables structure
* add new tables (user and auth)

+++ mcdrive/src/main/java/ru/karamoff/mcdrive/models/FoodpieceOrder.java
@@ -14,5 +14,5 @@ public class FoodpieceOrder {
     private Foodpiece foodpiece;
     private Boolean ready;
 
-    private Map<Modification, Integer> addedMods;
+    private Map<Modification, Short> addedMods;
 }
